requirements.txt
aiohttp
dnspython
colorama
beautifulsoup4
pyyaml
